Fix alignment's GetCurve for EdgeTX versions >= 2.9.0
In commit 8f3dd7f72e878607ab28af2b0c8ccca6e835fdd5[0] EdgeTX started
to emit the curve elements with the first element in the list having
the index 1. This is expected in LUA world but wasn't previously the
case.

SoarETX used to workaround the situation by rewriting the curve table.
With this commit the curve table is only rewritten when the curve
retrieved from EdgeTX does contain an element starting at position 0.

[0]: https://github.com/EdgeTX/edgetx/commit/8f3dd7f72e878607ab28af2b0c8ccca6e835fdd5